Move setup to server
